+++ packages/components/dropdown/src/use-dropdown.ts
@@ -1,8 +1,9 @@
 import type {PopoverProps} from "@nextui-org/popover";
 import type {MenuTriggerType} from "@react-types/menu";
 import type {Ref} from "react";
+import type {HTMLNextUIProps, PropGetter} from "@nextui-org/system";
 
-import {useProviderContext, type HTMLNextUIProps, type PropGetter} from "@nextui-org/system";
+import {useProviderContext} from "@nextui-org/system";
 import {useMenuTriggerState} from "@react-stately/menu";
 import {useMenuTrigger} from "@react-aria/menu";
 import {dropdown} from "@nextui-org/theme";
@@ -12,6 +13,7 @@ import {ariaShouldCloseOnInteractOutside} from "@nextui-org/aria-utils";
 import {useMemo, useRef} from "react";
 import {mergeProps} from "@react-aria/utils";
 import {MenuProps} from "@nextui-org/menu";
+import {CollectionElement} from "@react-types/shared";
 
 interface Props extends HTMLNextUIProps<"div"> {
   /**
@@ -40,6 +42,40 @@ interface Props extends HTMLNextUIProps<"div"> {
 }
 
 export type UseDropdownProps = Props & Omit<PopoverProps, "children" | "color" | "variant">;
+
+const getMenuItem = <T extends object>(props: Partial<MenuProps<T>> | undefined, key: string) => {
+  if (props) {
+    const mergedChildren = Array.isArray(props.children)
+      ? props.children
+      : [...(props?.items || [])];
+
+    if (mergedChildren && mergedChildren.length) {
+      const item = ((mergedChildren as CollectionElement<T>[]).find((item) => {
+        if (item.key === key) {
+          return item;
+        }
+      }) || {}) as {props: MenuProps};
+
+      return item;
+    }
+  }
+
+  return null;
+};
+
+const getCloseOnSelect = <T extends object>(
+  props: Partial<MenuProps<T>> | undefined,
+  key: string,
+  item?: any,
+) => {
+  const mergedItem = item || getMenuItem(props, key);
+
+  if (mergedItem && mergedItem.props && "closeOnSelect" in mergedItem.props) {
+    return mergedItem.props.closeOnSelect;
+  }
+
+  return props?.closeOnSelect;
+};
 
 export function useDropdown(props: UseDropdownProps) {
   const globalContext = useProviderContext();
@@ -151,7 +187,11 @@ export function useDropdown(props: UseDropdownProps) {
       menuProps,
       closeOnSelect,
       ...mergeProps(props, {
-        onAction: () => onMenuAction(props?.closeOnSelect),
+        onAction: (key: any, item?: any) => {
+          const closeOnSelect = getCloseOnSelect(props, key, item);
+
+          onMenuAction(closeOnSelect);
+        },
         onClose: state.close,
       }),
     } as MenuProps;