+++ packages/core/theme/src/components/autocomplete.ts
@@ -14,12 +14,16 @@ const autocomplete = tv({
       "translate-x-1",
       "cursor-text",
       "opacity-0",
+      "pointer-events-none",
       "text-default-500",
       "group-data-[invalid=true]:text-danger",
       "data-[visible=true]:opacity-100", // on mobile is always visible when there is a value
+      "data-[visible=true]:pointer-events-auto",
       "data-[visible=true]:cursor-pointer",
       "sm:data-[visible=true]:opacity-0", // only visible on hover
+      "sm:data-[visible=true]:pointer-events-none",
       "sm:group-data-[hover=true]:data-[visible=true]:opacity-100",
+      "sm:group-data-[hover=true]:data-[visible=true]:pointer-events-auto",
     ],
     selectorButton: "text-medium",
   },