+++ packages/hooks/use-ref-state/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nextui-org/use-ref-state
 
+## 2.0.7
+
+### Patch Changes
+
+- [#3607](https://github.com/nextui-org/nextui/pull/3607) [`eda316ab2`](https://github.com/nextui-org/nextui/commit/eda316ab24d390065931daf3aa816cbdbff4fb9a) Thanks [@AnYiEE](https://github.com/AnYiEE)! - ensure exposed functions wrap by `useCallback`
+
+## 2.0.6
+
+### Patch Changes
+
+- [#3512](https://github.com/nextui-org/nextui/pull/3512) [`2d2d300a1`](https://github.com/nextui-org/nextui/commit/2d2d300a12dbe20ca7ebd125daf3dce74efcbf34) Thanks [@wingkwong](https://github.com/wingkwong)! - fix conflicting versions in npm
+
 ## 2.0.5
 
 ### Patch Changes